+++ src/org/helioviewer/jhv/view/BaseView.java
@@ -53,7 +53,7 @@ public class BaseView implements View {
     }
 
     @Override
-    public void decode(Position viewpoint, double pixFactor, double factor) {
+    public void decode(Position viewpoint, double pixFactor, float factor) {
     }
 
     @Override
